SystemClockCoroutine.h: Fix bug that prevented working with CoroutineScheduler
The current version is not executed by `CoroutineScheduler`. This was difficult
to debug. The cause was traced to one of the 2 template parameters used to
instantiate `SystemClockCoroutine`. Both need to be some sort of
`ClockInterface`, and were set to `ace_time::hw::ClockInterface`. It turns out
that the second parameter must actually be `ace_routine::ClockInterface` to
allow the `CoroutineScheduler` to recognize the `SystemClockCoroutine`.